+++ app/src/main/java/com/gogreen/greenmachine/main/match/DrivingActivity.java
@@ -19,6 +19,7 @@ import android.widget.TimePicker;
 
 import com.gc.materialdesign.views.ButtonRectangle;
 import com.gogreen.greenmachine.R;
+import com.gogreen.greenmachine.interBack.InterBack;
 import com.gogreen.greenmachine.parseobjects.PrivateProfile;
 import com.parse.ParseException;
 import com.parse.ParseUser;
@@ -37,6 +38,8 @@ public class DrivingActivity extends ActionBarActivity {
     private EditText driverCarEditText;
     private EditText matchByEditText;
     private EditText arriveByEditText;
+
+    InterBack backend = new InterBack();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -197,16 +200,9 @@ public class DrivingActivity extends ActionBarActivity {
         return super.onOptionsItemSelected(item);
     }
 
-    private void setDefaultCar() {
-        ParseUser currUser = ParseUser.getCurrentUser();
-        PrivateProfile privProfile = (PrivateProfile) currUser.get("privateProfile");
-        try {
-            privProfile.fetchIfNeeded();
-        } catch(ParseException e) {
-            // Handle fetch fail
-        }
 
-        this.driverCarEditText.setText(privProfile.getDriverCar());
+    private void setDefaultCar() {
+        this.driverCarEditText.setText(backend.getDriverCarInfo());
     }
 
     private void startNextActivity() {
