+++ app/src/main/java/com/gogreen/greenmachine/main/match/DrivingHotspotSelectActivity.java
@@ -13,10 +13,10 @@ import android.widget.Toast;
 
 import com.gc.materialdesign.views.ButtonFloat;
 import com.gogreen.greenmachine.R;
+import com.gogreen.greenmachine.interBack.InterBack;
 import com.gogreen.greenmachine.parseobjects.Hotspot;
 import com.gogreen.greenmachine.parseobjects.MatchRoute;
 import com.gogreen.greenmachine.parseobjects.PublicProfile;
-import com.gogreen.greenmachine.util.Utils;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.location.LocationListener;
@@ -78,6 +78,7 @@ public class DrivingHotspotSelectActivity extends AppCompatActivity implements
     private Date arriveByDate;
     private MatchRoute.Destination destination;
     private Set<Hotspot> selectedHotspots;
+    private InterBack backend = new InterBack();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -95,7 +96,7 @@ public class DrivingHotspotSelectActivity extends AppCompatActivity implements
         this.mRequestingLocationUpdates = true;
 
         // Grab server hotspots
-        this.serverHotspots = getAllHotspots();
+        this.serverHotspots = backend.getAllHotspots();
 
         // Initialize selectedHotspots
         this.selectedHotspots = new HashSet<Hotspot>();
@@ -187,15 +188,7 @@ public class DrivingHotspotSelectActivity extends AppCompatActivity implements
         if (mCurrentLocation != null) {
             mLatitude = mCurrentLocation.getLatitude();
             mLongitude = mCurrentLocation.getLongitude();
-
-            // Fetch user's public profile
-            ParseUser currentUser = ParseUser.getCurrentUser();
-            PublicProfile pubProfile = (PublicProfile) currentUser.get("publicProfile");
-            Utils.getInstance().fetchParseObject(pubProfile);
-
-            // Insert coordinates into the user's public profile lastKnownLocation
-            ParseGeoPoint userLoc = new ParseGeoPoint(mLatitude, mLongitude);
-            pubProfile.setLastKnownLocation(userLoc);
+            backend.setUserLastKnownLocation(mLatitude, mLongitude);
         }
     }
 
@@ -291,8 +284,7 @@ public class DrivingHotspotSelectActivity extends AppCompatActivity implements
             Iterator iter = this.serverHotspots.iterator();
             while (iter.hasNext()) {
                 Hotspot h = (Hotspot) iter.next();
-                ParseGeoPoint parsePoint = h.getParseGeoPoint();
-                LatLng hotspotLoc = new LatLng(parsePoint.getLatitude(), parsePoint.getLongitude());
+                LatLng hotspotLoc = h.getHotspotLocation();
                 MarkerOptions markerOptions = new MarkerOptions().position(hotspotLoc)
                         .icon(BitmapDescriptorFactory.defaultMarker(30))
                         .title(h.getName())
@@ -324,24 +316,20 @@ public class DrivingHotspotSelectActivity extends AppCompatActivity implements
     public void setMarker(Marker m){
         m.setIcon(BitmapDescriptorFactory.defaultMarker(150));
         m.setAlpha(1.0f);
+
         // Grab location & add to Hotspot set
         LatLng mPoint = m.getPosition();
-        ParseGeoPoint hPoint = new ParseGeoPoint(mPoint.latitude, mPoint.longitude);
 
         // Find the original hotspot item and add it to the set
         Iterator iter = this.serverHotspots.iterator();
         while (iter.hasNext()) {
             Hotspot hSpot = (Hotspot) iter.next();
-            Utils.getInstance().fetchParseObject(hSpot);
-            if (isEqualParseGeoPoint(hPoint, hSpot.getParseGeoPoint())) {
+            backend.fetchIfNeeded(hSpot);
+            if (hSpot.isHotspotAtLocation(mPoint)) {
                 this.selectedHotspots.add(hSpot);
                 break;
             }
         }
-    }
-
-    private boolean isEqualParseGeoPoint(ParseGeoPoint p1, ParseGeoPoint p2) {
-        return (p1.getLatitude() == p2.getLatitude() && p1.getLongitude() == p2.getLongitude());
     }
 
     public void resetMarker(Marker m){
@@ -349,32 +337,17 @@ public class DrivingHotspotSelectActivity extends AppCompatActivity implements
         m.setIcon(BitmapDescriptorFactory.defaultMarker(30));
 
         LatLng mPoint = m.getPosition();
-        ParseGeoPoint hPoint = new ParseGeoPoint(mPoint.latitude, mPoint.longitude);
 
         // Find the original hotspot item and add it to the set
         Iterator iter = this.serverHotspots.iterator();
         while (iter.hasNext()) {
             Hotspot hSpot = (Hotspot) iter.next();
-            Utils.getInstance().fetchParseObject(hSpot);
-            if (isEqualParseGeoPoint(hPoint, hSpot.getParseGeoPoint())) {
+            backend.fetchIfNeeded(hSpot);
+            if (hSpot.isHotspotAtLocation(mPoint)) {
                 this.selectedHotspots.remove(hSpot);
                 break;
             }
         }
-    }
-
-    private Set<Hotspot> getAllHotspots() {
-        // Grab the hotspot set from the server
-        ParseQuery<Hotspot> hotspotQuery = ParseQuery.getQuery("Hotspot");
-        hotspotQuery.orderByDescending("hotspotId");
-        try {
-            serverHotspots = new HashSet<Hotspot>(hotspotQuery.find());
-        } catch (ParseException e) {
-            // Handle a server query fail
-            return null;
-        }
-
-        return serverHotspots;
     }
 
     private boolean checkForRiders() {
